% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{folderGetInfo}
\alias{folderGetInfo}
\alias{folderInit}
\title{Functions to support the Folder view}
\usage{
folderGetInfo(path, pattern, recursive, properties)

folderInit(folderInfo, files)
}
\arguments{
\item{path, pattern, recursive, properties}{arguments passed to
\link[base]{list.files}.}
}
\description{
Functions to support the Folder view
}

