% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R
\docType{import}
\name{reexports}
\alias{aceEditor}
\alias{dashboardBody}
\alias{dashboardHeader}
\alias{dashboardPage}
\alias{dashboardSidebar}
\alias{menuItem}
\alias{reexports}
\alias{renderDataTable}
\alias{sidebarMenu}
\alias{tabItems}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DT}{\code{\link[DT]{renderDataTable}}}

  \item{shinyAce}{\code{\link[shinyAce]{aceEditor}}}

  \item{shinydashboard}{\code{\link[shinydashboard]{dashboardHeader}}, \code{\link[shinydashboard]{dashboardSidebar}}, \code{\link[shinydashboard]{dashboardBody}}, \code{\link[shinydashboard]{dashboardPage}}, \code{\link[shinydashboard]{sidebarMenu}}, \code{\link[shinydashboard]{menuItem}}, \code{\link[shinydashboard]{tabItems}}}
}}
\keyword{internal}

