% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{settingsTabConfig}
\alias{settingsGetConfig}
\alias{settingsRestore}
\alias{settingsSave}
\alias{settingsTabConfig}
\title{Settings Elements}
\usage{
settingsTabConfig()

settingsSave(configText)

settingsRestore(session)

settingsGetConfig()
}
\arguments{
\item{configText}{(character)}

\item{session}{the current shiny session}
}
\description{
...
}

