% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{folderGet}
\alias{folderGet}
\title{Functions to support the Folder view}
\usage{
folderGet(path, pattern, recursive, properties)
}
\arguments{
\item{path, pattern, recursive, properties}{arguments passed to
\link[base]{list.files}.}
}
\description{
Functions to support the Folder view
}

