% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{boxWide}
\alias{boxFileContent}
\alias{boxWide}
\alias{tabFluid}
\title{UI Elements}
\usage{
boxWide(...)

tabFluid(tabName, ...)

boxFileContent(...)
}
\arguments{
\item{...}{something}

\item{tabName}{(character)}
}
\description{
Wrapper functions for UI-Elements.
}

